UART

+++ MCAL_Layer/mcal_std_types.h
@@ -17,10 +17,10 @@
 /* Section : Data Type Declarations */
 typedef unsigned char uint8;
 typedef unsigned short uint16;
-typedef unsigned int uint32;
+typedef unsigned long uint32;
 typedef signed char sint8;
 typedef signed short sint16;
-typedef signed int sint32;
+typedef signed long sint32;
 typedef uint8 Std_ReturnType;
 
 /* Section : Macro Declarations */
